backend_project/requirements.txt
Django<5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
requests==2.31.0
celery==5.4
redis==5.0.4
mysqlclient
django-sslserver # デプロイ時に削除